+++ jdaviz/utils.py
@@ -29,6 +29,14 @@ class SnackbarQueue:
         self.first = True
 
     def put(self, state, msg):
+        if msg.color not in ['info', 'warning', 'error', 'success', None]:
+            raise ValueError(f"color ({msg.color}) must be on of: info, warning, error, success")
+
+        if not msg.loading:
+            now = time.localtime()
+            timestamp = f'{now.tm_hour}:{now.tm_min:02d}:{now.tm_sec:02d}'
+            state.snackbar_history.append({'time': timestamp, 'text': msg.text, 'color': msg.color})
+
         if msg.loading:
             # immediately show the loading message indefinitely until cleared by a new message
             # with loading=False (or overwritten by a new indefinite message with loading=True)
@@ -39,7 +47,24 @@ class SnackbarQueue:
             self.loading = False
             self._write_message(state, msg)
         else:
-            self.queue.append(msg)
+            if msg.color in ['warning', 'error']:
+                if (state.snackbar.get('show') and
+                        ((msg.color == 'warning' and state.snackbar.get('color') in ['warning', 'error']) or  # noqa
+                         (msg.color == 'error' and state.snackbar.get('color') in ['error']))):
+                    # put this NEXT in the queue immediately FOLLOWING all warning/errors
+                    non_warning_error = [msg.color not in ['warning', 'error'] for msg in self.queue]  # noqa
+                    if True in non_warning_error:
+                        # insert BEFORE index
+                        self.queue.insert(non_warning_error.index(True), msg)
+                    else:
+                        self.queue.append(msg)
+                else:
+                    # interrupt the queue IMMEDIATELY
+                    # (any currently shown messages will repeat after)
+                    self._write_message(state, msg)
+            else:
+                # put this LAST in the queue
+                self.queue.append(msg)
             if len(self.queue) == 1:
                 self._write_message(state, msg)
 
@@ -70,7 +95,6 @@ class SnackbarQueue:
             self._write_message(state, msg)
 
     def _write_message(self, state, msg):
-
         state.snackbar['show'] = False
         state.snackbar['text'] = msg.text
         state.snackbar['color'] = msg.color
@@ -102,13 +126,15 @@ class SnackbarQueue:
 
         # create the timeout function which will close this message and
         # show the next message if one has been added to the queue since
-        def sleep_function(timeout):
+        def sleep_function(timeout, text):
             timeout_ = float(timeout) / 1000
             time.sleep(timeout_)
-            self.close_current_message(state)
+            if state.snackbar['show'] and state.snackbar['text'] == text:
+                # don't close the next message if the user manually clicked close!
+                self.close_current_message(state)
 
         x = threading.Thread(target=sleep_function,
-                             args=(timeout,),
+                             args=(timeout, msg.text),
                              daemon=True)
         x.start()